--- test/langtools/tools/javac/valhalla/primitive-classes/ConflictingSuperInterfaceTest.out
@@ -1,2 +1,0 @@
-ConflictingSuperInterfaceTest.java:12:22: compiler.err.cant.inherit.diff.arg: ConflictingSuperInterfaceTest.I, java.lang.Integer, java.lang.String
-1 error
